requirements.txt
asgiref==3.5.2
Django==4.0.4
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pytz==2022.7.1
sqlparse==0.4.2
tzdata==2022.7
whitenoise==6.1.0
